+++ docs/conf.py
@@ -16,15 +16,15 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath(
-    os.path.join('..', 'src', 'nanocurrency')
+    os.path.join('..', 'src', 'nanolib')
 ))
 
-import nanocurrency
+import nanolib
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'pynanocurrency'
+project = 'nanolib'
 copyright = '2019, Janne Pulkkinen'
 author = 'Janne Pulkkinen'
 
@@ -32,9 +32,9 @@ author = 'Janne Pulkkinen'
 html_show_copyright = False
 
 # The short X.Y version
-version = nanocurrency.__version__
+version = nanolib.__version__
 # The full version, including alpha/beta/rc tags
-release = nanocurrency.__version__
+release = nanolib.__version__
 
 
 # -- General configuration ---------------------------------------------------
@@ -113,7 +113,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'pynanocurrencydoc'
+htmlhelp_basename = 'nanolibdoc'
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -140,7 +140,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'pynanocurrency.tex', 'pynanocurrency Documentation',
+    (master_doc, 'nanolib.tex', 'nanolib Documentation',
      'Janne Pulkkinen', 'manual'),
 ]
 
@@ -150,7 +150,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'pynanocurrency', 'pynanocurrency Documentation',
+    (master_doc, 'nanolib', 'nanolib Documentation',
      [author], 1)
 ]
 
@@ -161,8 +161,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'pynanocurrency', 'pynanocurrency Documentation',
-     author, 'pynanocurrency', 'Utilities for working with NANO cryptocurrency protocol',
+    (master_doc, 'nanolib', 'nanolib Documentation',
+     author, 'nanolib', 'Utilities for working with NANO cryptocurrency protocol',
      'Miscellaneous'),
 ]