+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from nanocurrency.blocks import Block
+from nanolib.blocks import Block
 
 from tests.data import BLOCKS
 
@@ -9,9 +9,9 @@ from tests.data import BLOCKS
 def low_pow_threshold(monkeypatch):
     # Use a far lower default threshold for unit tests
     TEST_THRESHOLD = 9459044173002835106
-    monkeypatch.setattr("nanocurrency.blocks.WORK_THRESHOLD", TEST_THRESHOLD)
-    monkeypatch.setattr("nanocurrency.work.WORK_THRESHOLD", TEST_THRESHOLD)
-    monkeypatch.setattr("nanocurrency.WORK_THRESHOLD", TEST_THRESHOLD)
+    monkeypatch.setattr("nanolib.blocks.WORK_THRESHOLD", TEST_THRESHOLD)
+    monkeypatch.setattr("nanolib.work.WORK_THRESHOLD", TEST_THRESHOLD)
+    monkeypatch.setattr("nanolib.WORK_THRESHOLD", TEST_THRESHOLD)
 
 
 @pytest.fixture(scope="function")