+++ tests/test_accounts.py
@@ -1,14 +1,14 @@
 import pytest
 
-from nanocurrency.exceptions import (
+from nanolib.exceptions import (
     InvalidSeed, InvalidPrivateKey, InvalidPublicKey, InvalidAccount)
-from nanocurrency.accounts import (
+from nanolib.accounts import (
     AccountIDPrefix, get_account_key_pair, get_account_id,
     get_account_public_key, generate_account_private_key,
     generate_account_key_pair, generate_account_id, generate_seed,
     is_account_id_valid
 )
-from nanocurrency.util import is_hex
+from nanolib.util import is_hex
 
 
 SEED = "bba817a4fa1418e10d014c99055c4922afa0f84b324e7850baf4b3b8b6af1a5b"
@@ -74,7 +74,7 @@ def test_generate_account_key_pair():
 
 def test_get_account_id():
     """
-    Test nanocurrency.accounts.get_account_id
+    Test nanolib.accounts.get_account_id
     """
     # From PUBLIC KEY
     with pytest.raises(InvalidPublicKey):
@@ -126,7 +126,7 @@ def test_get_account_id():
 
 def test_get_account_public_key():
     """
-    Test nanocurrency.accounts.get_account_public_key
+    Test nanolib.accounts.get_account_public_key
     """
     # From ACCOUNT ID
     with pytest.raises(InvalidAccount):